+++ processes/__init__.py
@@ -16,3 +16,4 @@ from .migration import MigrationProcess
 from .dbMaintenance import DatabaseMaintenanceProcess
 from .chicagoISAC import ChicagoISACProcess
 from .UofSC import UofSCProcess
+from .loc import LOCProcess